fix repeat call get user info api issue
