+++ static/js/controller/auth.js
@@ -101,6 +101,54 @@ export async function checkUserState() {
   }
 }
 
+export async function fetchAndStoreUserInfo() {
+  const token = localStorage.getItem("userToken");
+  try {
+    const response = await fetch(userSignInUrl, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    if (!response.ok) {
+      throw new Error("Failed to fetch user info");
+    }
+
+    if (data && data.data) {
+      localStorage.setItem("userName", data.data.name);
+    }
+  } catch (error) {
+    console.error("Error fetching user info:", error);
+  }
+}
+
+export async function fetchGetUserName() {
+  const token = localStorage.getItem("userToken");
+  try {
+    const response = await fetch(userSignInUrl, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+
+    if (!response.ok) {
+      console.error("Failed to fetch booking details:", response.status);
+      return;
+    }
+
+    const data = await response.json();
+
+    if (!data || !data.data) {
+      console.error("No booking data available");
+      return;
+    }
+    console.log(data.data.name);
+    localStorage.setItem("userName", data.data.name);
+    return data.data.name;
+  } catch (error) {
+    console.error("Error fetching attraction:", error);
+  }
+}
+
 export function setupLogoutListener() {
   const logout = document.querySelector(".logout");
   logout.addEventListener("click", logOut);
@@ -116,4 +164,5 @@ function logOut() {
 export function initialize() {
   setupEventListeners();
   setupLogoutListener();
+  fetchGetUserName();
 }
